# apps/api/src/tasks/dto/delete-task-response.dto.ts
import { z } from 'zod';
import { createZodDto } from 'nestjs-zod';

const DeleteTaskResponseSchema = z.object({
  success: z.boolean(),
});

export class DeleteTaskResponseDto extends createZodDto(
  DeleteTaskResponseSchema,
) {}